Basic device tracking support
Adds support for keeping a history of accelerometer samples, applying
some basic statistical methods to them (outlier culling), and computing
mean and trimmed mean acceleration.

This commit also removes the libm crate. It seems it's un(der)maintained
and full of overflow bugs. We're only using one function from it at the
moment (sqrt) so we can just implement that for now.